Rename remaining headers
