+++ src/gtt_file_io.c
@@ -28,7 +28,7 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "app.h"
+#include "gtt_application_window.h"
 #include "gtt_current_project.h"
 #include "gtt_err_throw.h"
 #include "gtt_gsettings_io.h"
