+++ src/gtt_menu_commands.c
@@ -24,7 +24,7 @@
 #include <gnome.h>
 #include <string.h>
 
-#include "app.h"
+#include "gtt_application_window.h"
 #include "gtt_current_project.h"
 #include "gtt_err_throw.h"
 #include "gtt_file_io.h"
