+++ src/gtt_preferences.c
@@ -24,7 +24,7 @@
 #include <qof.h>
 #include <string.h>
 
-#include "app.h"
+#include "gtt_application_window.h"
 #include "gtt_current_project.h"
 #include "gtt_help_popup.h"
 #include "gtt.h"
