+++ src/gtt_timer.c
@@ -25,7 +25,7 @@
 #include <string.h>
 
 #include "gtt_activation_dialog.h"
-#include "app.h"
+#include "gtt_application_window.h"
 #include "gtt_current_project.h"
 #include "gtt.h"
 #include "gtt_idle_dialog.h"
